app/pyproject.toml
[project]
name = "devops-monitoring-app"
version = "0.1.0"
description = "Laptop monitoring app for DevOps module"
dependencies = [
    "psutil>=6.0.0",
    "paho-mqtt>=2.0.0",
]

[tool.uv]
package = false
